Rename Folders and paths
